pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Pau <plabartabajo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
ta = "^0.10.2"
bytewax = "^0.15.1"
websocket-client = "^1.5.1"
pyOpenSSL = "^23.1.1"
hopsworks = "3.2.0"
streamlit = "^1.24.0"
bokeh = "2.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"
pip-tools = "^6.12.3"
pylint = "^2.17.1"
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
]
